+++ src/routes/index.ts
@@ -1,6 +1,0 @@
-export * from './error/errorRoutes.js';
-export * from './healthCheckRoute.js';
-export * from './infoRoute.js';
-export * from './mainRoute.js';
-export * from './packagesRoute.js';
-export * from './swaggerRoute.js';