+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# [1.11.0](https://github.com/ffflorian/pkgsource/compare/v1.10.0...v1.11.0) (2026-04-13)
+
+
+### Features
+
+* Remove vercel config and use SOURCE_COMMIT from Coolify ([21006ff](https://github.com/ffflorian/pkgsource/commit/21006ff23e51a84ddf0f6eb24e63bc4032631d30))
+
+# [1.10.0](https://github.com/ffflorian/pkgsource/compare/v1.9.0...v1.10.0) (2026-04-07)
+
+
+### Features
+
+* Add logging to exception filter ([#1773](https://github.com/ffflorian/pkgsource/issues/1773)) ([8694f30](https://github.com/ffflorian/pkgsource/commit/8694f3007bc1250d5321fad32a268105254421fe))
+
+# [1.9.0](https://github.com/ffflorian/pkgsource/compare/v1.8.2...v1.9.0) (2026-04-07)
+
+
+### Features
+
+* Replace Express with NestJS, remove pm2 ([#1771](https://github.com/ffflorian/pkgsource/issues/1771)) ([49d7c7b](https://github.com/ffflorian/pkgsource/commit/49d7c7b9eee504d6702a60d43a47cd330d445c07))
+
 ## [1.8.2](https://github.com/ffflorian/pkgsource/compare/v1.8.1...v1.8.2) (2026-02-05)